HVR-16 Added functionality to upload images for avatars

+++ VeterinaryCustomer.Web/Controllers/AvatarController.cs
@@ -1,9 +1,12 @@
+using System;
+using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using VeterinaryCustomer.Domain.Models;
 using VeterinaryCustomer.Repositories.Repositories;
+using VeterinaryCustomer.Services.Aws;
 using VeterinaryCustomer.Web.Attributes;
 
 namespace VeterinaryCustomer.Web.Controllers
@@ -18,13 +21,16 @@ namespace VeterinaryCustomer.Web.Controllers
 
         private readonly IAvatarRepository _avatarRepository;
 
+        private readonly IS3Service _s3Service;
+
         #endregion
 
         #region snippet_Constructors
 
-        public AvatarController(IAvatarRepository avatarRepository)
+        public AvatarController(IAvatarRepository avatarRepository, IS3Service s3Service)
         {
             _avatarRepository = avatarRepository;
+            _s3Service = s3Service;
         }
 
         #endregion
@@ -38,6 +44,11 @@ namespace VeterinaryCustomer.Web.Controllers
 
             if (avatar is null) return NotFound();
 
+            var bucketName = Environment.GetEnvironmentVariable("S3_BUCKET");
+            var s3Endpoint = Environment.GetEnvironmentVariable("S3_ENDPOINT");
+
+            avatar.Path = $"{s3Endpoint}/{bucketName}/{avatar.Path}";
+
             return Ok(avatar);
         }
 
@@ -49,13 +60,16 @@ namespace VeterinaryCustomer.Web.Controllers
             [Required] [FromForm] IFormFile image
         )
         {
+            var imageStream = image.OpenReadStream();
+            var fullPath = await _s3Service.PutObjectAsync(image.FileName, userId, imageStream);
+
             var avatar = new Avatar
             {
                 CustomerId = userId,
                 Path = $"{userId}/{image.FileName}"
             };
-
             await _avatarRepository.CreateAsync(avatar);
+            avatar.Path = fullPath;
 
             return Created("", avatar);
         }
@@ -72,9 +86,12 @@ namespace VeterinaryCustomer.Web.Controllers
 
             if (avatar is null) return NotFound();
 
-            avatar.Path = $"{userId}/{image.FileName}";
+            var imageStream = image.OpenReadStream();
+            var fullPath = await _s3Service.PutObjectAsync(image.FileName, userId, imageStream);
 
+            avatar.Path = $"{userId}/{image.FileName}";
             await _avatarRepository.UpdateAsync(avatar);
+            avatar.Path = fullPath;
 
             return Ok(avatar);
         }
@@ -83,7 +100,17 @@ namespace VeterinaryCustomer.Web.Controllers
         [Avatar]
         public async Task<IActionResult> DeleteMeAsync([FromHeader(Name = "user-id")] string userId)
         {
-            await _avatarRepository.DeleteAsync(userId);
+            var avatar = await _avatarRepository.GetByCustomerIdAsync(userId);
+
+            if (avatar is null) return NotFound();
+
+            IEnumerable<Task> tasks = new List<Task>
+            {
+                _s3Service.DeleteObjectAsync(avatar.Path),
+                _avatarRepository.DeleteAsync(userId)
+            };
+            await Task.WhenAll(tasks);
+
             return NoContent();
         }
